+++ src/components/HURUmap/ChartFilter/index.js
@@ -3,13 +3,14 @@ import React from "react";
 
 import useStyles from "./useStyles";
 
-import Input from "@/pesayetu/components/Select";
+import Input from "@/pesayetu/components/Input";
 
 const ChartFilter = ({ label, options, selected, ...props }) => {
   const classes = useStyles(props);
   return (
     <div className={classes.root}>
       <Input
+        helperText="select a value"
         label={label}
         classes={{ select: classes.select }}
         options={options}
