+++ src/components/Input/index.stories.js
